[FINDBUGS] First bunch of fixes for problems reported by FindBugs in drools-core (#1971)
* [findbugs] Create classloaders only inside doPrivileged block

* [findbugs] Use correct map for accumulate imports initialization

* [findbugs] Uncallable method defined in anonymous class

* [findbugs] Impossible cast

* [findbugs] Code checks for specific values returned by compareTo

* [findbugs] Bad attempt to compute abs. value of signed 32-bit hashcode

* [findbugs] Nullcheck of value previously dereferenced

* [findbugs] Uninit. read of field method called from constr. of superclass

* [findbugs] equals overrides equals in superclass and may not be symmetric

* [findbugs] Possible null pointer dereference

* [findbugs] Fix failing tests

* [findbugs] Should be a static inner class

* [findbugs] Private method is never called

* [findbugs] Inefficient use of keySet iterator instead of entrySet iterator

* [findbugs] Private method is never called

* [findbugs] Method concatenates strings using + in a loop

* [findbugs] Method invokes inefficient Number constructor

* [findbugs] Boxing/unboxing to parse a primitive

* [findbugs] Boxed value is unboxed and then immediately reboxed

* [findbugs] Other boxing/unboxing issues

* [findbugs] Method allocates a boxed primitive just to call toString

* [findbugs] Fix test

* [findbugs] Sequence of calls to concurrent abstraction may not be atomic

* [findbugs] Possible double check of field (double-checked locking)

* [findbugs] Method does not release lock on all exception paths

* [findbugs] Class is Serializable, but doesn't define serialVersionUID

* [findbugs] Method ignores exceptional return value

* [findbugs] Method uses the same code for two branches

* [findbugs] Write to static field from instance method

* Revert to instanceOf comparison

+++ kogito-drools/drools-core/src/main/java/org/drools/core/base/ClassFieldAccessorFactory.java
@@ -109,7 +109,7 @@ public class ClassFieldAccessorFactory {
                     }
                 }
                 if ( fieldType != null && getterMethod != null ) {
-                    final String className = ClassFieldAccessorFactory.BASE_PACKAGE + "/" + Type.getInternalName( clazz ) + Math.abs( System.identityHashCode( clazz ) ) + "$" + getterMethod.getName();
+                    final String className = ClassFieldAccessorFactory.BASE_PACKAGE + "/" + Type.getInternalName( clazz ) + Math.abs( (long) System.identityHashCode( clazz ) ) + "$" + getterMethod.getName();
 
                     // generating byte array to create target class
                     final byte[] bytes = dumpReader( clazz,
@@ -185,7 +185,7 @@ public class ClassFieldAccessorFactory {
             }
             if ( setterMethod != null ) {
                 final Class< ? > fieldType = setterMethod.getParameterTypes()[0];
-                final String className = ClassFieldAccessorFactory.BASE_PACKAGE + "/" + Type.getInternalName( clazz ) + Math.abs( System.identityHashCode( clazz ) ) + "$" + setterMethod.getName();
+                final String className = ClassFieldAccessorFactory.BASE_PACKAGE + "/" + Type.getInternalName( clazz ) + Math.abs( (long) System.identityHashCode( clazz ) ) + "$" + setterMethod.getName();
 
                 // generating byte array to create target class
                 final byte[] bytes = dumpWriter( clazz,
